third_party/WebKit/LayoutTests/fast/forms/input-set-selectionStart-set-selectionEnd-expected.txt

Test for BUG=367736: Blink does not respect input.selectionStart and input.selectionEnd for some cases.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstInput.selectionStart is 0
PASS firstInput.selectionEnd is 0
PASS firstInput.selectionStart is 3
PASS firstInput.selectionEnd is 5
PASS secondInput.selectionStart is 1
PASS secondInput.selectionEnd is 1
PASS secondInput.selectionStart is 1
PASS secondInput.selectionEnd is 1
PASS successfullyParsed is true

TEST COMPLETE
 
